update Assignment 1 (edits 1 and 2)
I forgot to commit my edit when I answered the first 2 questions x.x

I grinded out the assignement for the past 4 hours (6 PM - 10 PM) and I was able to finish up to question 5 of the assignment.  Hopefully the rest of the day tomorrow is enought time to finish the assignment before I submit it.