update Assignment 2 (final edit)
I spent the whole day trying to finish the assignment because I had put it off for so long. I could have tried to work on it a little more but I didn't because why am I like this.

Additionally I added more comments to other parts of the assignment.

I also converted the assignment to pdf formed and organized it in the folder accordingly.